updated filter.go
