+++ internal/filter/formatfilter.go
@@ -1,4 +1,4 @@
-package main
+package filter
 
 import (
 	"fmt"
