+++ internal/filter/utils.go
@@ -1,4 +1,4 @@
-package main
+package filter
 
 import (
 	"context"
